+++ microrts_trp/__init__.py
@@ -6,11 +6,13 @@ Implement the same calculation as the official tournament result analysis.
 
 __version__ = "0.1.1"
 
-import click
-import pandas as pd
-from tabulate import tabulate
+from pathlib import Path
+from dataclasses import dataclass
+from typing import Optional
+import functools
 
-from microrts_trp.tournament import parse_map_folder
+import numpy as np
+import pandas as pd
 
 
 def shorten_bot_name(name: str, max_length=35) -> str:
@@ -19,195 +21,191 @@ def shorten_bot_name(name: str, max_length=35) -> str:
     else:
         return name
 
+@dataclass
+class TournamentConfig:
+    """Configuration for each run of the tournament."""
 
-@click.group()
-def cli():
-    pass
-
-
-@cli.command(short_help="Show this message.")
-@click.pass_context
-def help(ctx):
-    click.echo(cli.get_help(ctx))
+    iterations: int
+    max_game_length: int
+    time_budget: int
 
 
-@cli.command(name="view", short_help="Parse a map-folder and show the win rates table.")
-@click.argument("map_folder")
-@click.option(
-    "--detail/--no-detail",
-    show_default=True,
-    default=False,
-    help="Show the detailed win rates table or show only the win rates.",
-)
-@click.option("--full-bot-name", help="Do not shorten the name of the bot.")
-@click.option(
-    "--format",
-    type=click.Choice(["github", "latex"]),
-    show_default=True,
-    default="github",
-    help="Do not shorten the name of the bot.",
-)
-def view_map_folder_command(map_folder, detail, full_bot_name, format):
-    map_result = parse_map_folder(map_folder)
-    win_rates = map_result.format_win_rates_for_human(detail, full_bot_name)
+@dataclass
+class TournamentResult:
+    """Representation of the tournament.csv file."""
 
-    print(f"Map: {map_result.name}")
-    print(f"Number of AIs: {map_result.win_rates.shape[0]}")
-    print(
-        f"Total battles / bot: {len(map_result.tournaments[0].ai_names) * map_result.total_iterations}"
-    )
-    print(
-        tabulate(
-            win_rates,
-            tablefmt=format,
-            headers=["Bot"] + win_rates.columns.tolist(),
-            floatfmt=".2f",
+    ai_names: list[str]
+    maps: list[str]
+    config: TournamentConfig
+    wins_table: pd.DataFrame
+    ties_table: pd.DataFrame
+
+    def calculate_scores(self):
+        return self.wins_table + (self.ties_table * 0.5)
+
+
+@dataclass
+class MapResult:
+    name: str
+    tournaments: list[TournamentResult]
+
+    def __str__(self):
+        return f"MapResult[name={self.name}, win_rates={self.win_rates}]"
+
+    @functools.cached_property
+    def total_iterations(self) -> int:
+        total_iterations = 0
+        for t in self.tournaments:
+            total_iterations += t.config.iterations
+        return total_iterations
+
+    @functools.cached_property
+    def final_scores(self) -> pd.DataFrame:
+        final_scores = None
+
+        for t in self.tournaments:
+            if final_scores is None:
+                final_scores = t.calculate_scores()
+            else:
+                final_scores += t.calculate_scores()
+        return final_scores
+
+    @functools.cached_property
+    def win_rates(self) -> pd.DataFrame:
+        win_rates_table = (self.final_scores * 100.0) / self.total_iterations
+
+        return win_rates_table
+    
+    def format_win_rates_for_human(self, show_detail=False, show_full_bot_name=False) -> pd.DataFrame:
+        final_scores = self.final_scores.copy()
+        final_scores["Total"] = final_scores.sum(axis=1)
+
+        win_rates = self.win_rates.copy()
+        win_rates["Win Rate"] = (
+            final_scores["Total"]
+            * 100.0
+            / (len(self.tournaments[0].ai_names) * self.total_iterations)
         )
-    )
+
+        # Move 'Win Rate' column to the left.
+        cols = win_rates.columns.tolist()
+        cols = cols[-1:] + cols[:-1]
+        win_rates = win_rates.loc[:, cols]
+
+        win_rates = win_rates.sort_values(by="Win Rate", ascending=False)
+
+        if not show_detail:
+            # Remove the detail columns.
+            cols = win_rates.columns.tolist()
+            win_rates = win_rates.drop(columns=cols[1:])
+
+        if not show_full_bot_name:
+            # Rename the bots in the index.
+            win_rates.index = win_rates.index.map(shorten_bot_name)
+
+            # Rename the bots in the column's name.
+            col_names = win_rates.columns.values.tolist()
+            new_col_names = {name: shorten_bot_name(name) for name in col_names}
+            win_rates = win_rates.rename(columns=new_col_names)
+
+        return win_rates
 
 
-@cli.command(name="compare", short_help="Compare win rates of two map folders.")
-@click.argument("map_folder_1")
-@click.argument("map_folder_2")
-@click.option("--full-bot-name", help="Do not shorten the name of the bot.")
-@click.option(
-    "--format",
-    type=click.Choice(["github", "latex"]),
-    show_default=True,
-    default="github",
-    help="Do not shorten the name of the bot.",
-)
-def compare_map_folders_command(map_folder_1, map_folder_2, full_bot_name, format):
-    map_result_1 = parse_map_folder(map_folder_1)
-    map_result_2 = parse_map_folder(map_folder_2)
+def parse_tournament_file(path: Path | str) -> Optional[TournamentResult]:
+    if isinstance(path, str):
+        path = Path(path)
 
-    # Sanity checks.
+    with open(path, "r") as in_f:
+        try:
+            lines = in_f.readlines()
 
-    win_rates_1 = map_result_1.format_win_rates_for_human(
-        show_detail=False, show_full_bot_name=full_bot_name
-    )
-    win_rates_2 = map_result_2.format_win_rates_for_human(
-        show_detail=False, show_full_bot_name=full_bot_name
-    )
+            current_section: str = ""
+            ai_names: list[str] = []
+            maps: list[str] = []
+            config = dict()
+            wins_table: Optional[np.ndarray] = None
+            ties_table: Optional[np.ndarray] = None
+            row_count: int = 0
+            for line in lines:
+                # Capture the section change.
+                match line.strip():
+                    case "AIs":
+                        current_section = "ai-list"
+                        continue
+                    case "maps":
+                        current_section = "map-list"
+                        continue
+                    case "Wins:":
+                        current_section = "wins-table"
+                        wins_table = np.zeros((len(ai_names), len(ai_names)))
+                        row_count = 0
+                        continue
+                    case "Ties:":
+                        current_section = "ties-table"
+                        ties_table = np.zeros((len(ai_names), len(ai_names)))
+                        row_count = 0
+                        continue
+                    case "Average Game Length:":
+                        current_section = "avg-game-length-table"
+                        continue
 
-    df = pd.DataFrame()
-    df[map_result_1.name] = win_rates_1["Win Rate"]
-    df[map_result_2.name] = win_rates_2["Win Rate"]
-    changes = df[map_result_2.name] - df[map_result_1.name]
-    df["Change (%)"] = (changes / df[map_result_1.name]) * 100.0
-    df = df.sort_values(by="Change (%)", ascending=False)
+                tokens = line.split("\t")
+                if len(tokens) > 1:
+                    if tokens[0] == "iterations":
+                        config["iterations"] = int(tokens[1])
+                    elif len(tokens) > 1 and tokens[0] == "maxGameLength":
+                        config["max_game_length"] = int(tokens[1])
+                    elif len(tokens) > 1 and tokens[0] == "timeBudget":
+                        config["time_budget"] = int(tokens[1])
 
-    print(
-        tabulate(
-            df,
-            tablefmt=format,
-            headers=["Bot"] + df.columns.tolist(),
-            floatfmt=".2f",
-        )
-    )
+                match current_section:
+                    case "map-list":
+                        maps.append(line.strip())
+                    case "ai-list":
+                        ai_names.append(line.strip())
+                    case "wins-table":
+                        wins = [
+                            int(t)
+                            for t in line.strip().split("\t")
+                            if len(t.strip()) != 0
+                        ]
+                        wins_table[row_count, :] = wins
+                        row_count += 1
+                    case "ties-table":
+                        tokens = [
+                            int(t)
+                            for t in line.strip().split("\t")
+                            if len(t.strip()) != 0
+                        ]
+                        ties_table[row_count, :] = tokens
+                        row_count += 1
+
+            return TournamentResult(
+                ai_names,
+                maps,
+                TournamentConfig(**config),
+                pd.DataFrame(data=wins_table, columns=ai_names, index=ai_names),
+                pd.DataFrame(data=ties_table, columns=ai_names, index=ai_names),
+            )
+        except Exception:
+            return None
 
 
-@cli.command(
-    name="focus",
-    short_help="Show focused result of one bot. How well it performs against the other bots \
-                and how well the other bot perform against it.",
-)
-@click.argument("map_folder")
-@click.argument("bot_name")
-@click.option("--full-bot-name", help="Do not shorten the name of the bot.")
-@click.option(
-    "--format",
-    type=click.Choice(["github", "latex"]),
-    show_default=True,
-    default="github",
-    help="Do not shorten the name of the bot.",
-)
-def focus_command(map_folder, bot_name, full_bot_name, format):
-    map_result = parse_map_folder(map_folder)
+def parse_map_folder(path: Path | str) -> MapResult:
+    if isinstance(path, str):
+        path = Path(path)
 
-    target_bot_vs_other = pd.DataFrame()
-    target_bot_vs_other[bot_name] = map_result.win_rates.loc[bot_name].transpose()
-    # Remove itself from the table.
-    target_bot_vs_other = target_bot_vs_other.drop(bot_name)
-    target_bot_vs_other = target_bot_vs_other.sort_values(by=bot_name, ascending=False)
+    path = path.absolute()
 
+    if not path.exists():
+        raise RuntimeError(f"path '{str(path)}' does not exist")
 
-    print(f"Map: {map_result.name}")
-    print(f"Number of AIs: {map_result.win_rates.shape[0]}")
-    print(f"Bot of Interest: {bot_name}")
-    print(f"\nHow well {bot_name} performed against the other bots:")
-    print(
-        tabulate(
-            target_bot_vs_other,
-            tablefmt=format,
-            headers=[bot_name],
-            floatfmt=".2f",
-        )
-    )
+    if path.is_file():
+        # Assume that the tournament.csv is viewed directly.
+        raw_tournament_files = [path]
+    elif path.is_dir():
+        raw_tournament_files = list(path.glob("*/tournament.csv"))
 
-    other_vs_target_bot = pd.DataFrame()
-    other_vs_target_bot[bot_name] = map_result.win_rates[bot_name]
-    other_vs_target_bot = other_vs_target_bot.drop(bot_name)
-    other_vs_target_bot = other_vs_target_bot.sort_values(by=bot_name, ascending=False)
-    print(f"\nHow well the other bots performed against {bot_name}:")
-    print(
-        tabulate(
-            other_vs_target_bot,
-            tablefmt=format,
-            headers=["Other bots", bot_name],
-            floatfmt=".2f",
-        )
-    )
-
-@cli.command(
-    name="focus-compare",
-    short_help="Show focused result of one bot. How well it performs against the other bots \
-                and how well the other bot perform against it.",
-)
-@click.argument("map_folder_1")
-@click.argument("map_folder_2")
-@click.argument("bot_name")
-@click.option("--full-bot-name", help="Do not shorten the name of the bot.")
-@click.option(
-    "--format",
-    type=click.Choice(["github", "latex"]),
-    show_default=True,
-    default="github",
-    help="Do not shorten the name of the bot.",
-)
-def focus_compare_command(map_folder_1, map_folder_2, bot_name, full_bot_name, format):
-    map_result_1 = parse_map_folder(map_folder_1)
-    map_result_2 = parse_map_folder(map_folder_2)
-
-    target_bot_vs_other = pd.DataFrame()
-    target_bot_vs_other[map_result_1.name] = map_result_1.win_rates.loc[bot_name].transpose()
-    target_bot_vs_other[map_result_2.name] = map_result_2.win_rates.loc[bot_name].transpose()
-    # Remove itself from the table.
-    target_bot_vs_other = target_bot_vs_other.drop(bot_name)
-    target_bot_vs_other = target_bot_vs_other.sort_values(by=map_result_2.name, ascending=False)
-
-    print(f"Bot of Interest: {bot_name}")
-    print(f"\nHow well {bot_name} performed against the other bots:")
-    print(
-        tabulate(
-            target_bot_vs_other,
-            tablefmt=format,
-            headers=[map_result_1.name, map_result_2.name],
-            floatfmt=".2f",
-        )
-    )
-
-    other_vs_target_bot = pd.DataFrame()
-    other_vs_target_bot[map_result_1.name] = map_result_1.win_rates[bot_name]
-    other_vs_target_bot[map_result_2.name] = map_result_2.win_rates[bot_name]
-    other_vs_target_bot = other_vs_target_bot.drop(bot_name)
-    other_vs_target_bot = other_vs_target_bot.sort_values(by=map_result_2.name, ascending=False)
-    print(f"\nHow well the other bots performed against {bot_name}:")
-    print(
-        tabulate(
-            other_vs_target_bot,
-            tablefmt=format,
-            headers=["Other bots", map_result_1.name, map_result_2.name],
-            floatfmt=".2f",
-        )
-    )
+    tournaments = [parse_tournament_file(f) for f in raw_tournament_files]
+    map_result = MapResult(name=path.name, tournaments=tournaments)
+    return map_result